pyproject.toml
[tool.poetry]
name = "azurechatgpt"
version = "0.1.0"
description = ""
authors = ["Ciprian Turcu <ciprian.turcu@globallogic.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.29.0"
openai = "0.28.0"
mypy = "^1.8.0"
python-dotenv = "^1.0.0"
pypdf = "^4.0.0"
langchain = "^0.1.2"
chromadb = "^0.4.22"
sentence-transformers = "^2.2.2"
umap = "^0.1.1"
torch = "^2.2.1"
tqdm = "^4.66.2"
matplotlib = "^3.8.3"
pdfminer = "^20191125"
tabula-py = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
